Refactor `alloc/rc.rs` into separate files
- `alloc::rc::Rc` now lives in `libraries/alloc/rc/rc.rs`
- `alloc::rc::Weak` now lives in `libraries/alloc/rc/weak.rs`
-  the majority of tests now live in `alloc/tests/arc.rs`
-  one test which tests internals is in `alloc/rc/rc.rs`

Other changes:

- Remove trailing newline
- Remove unused re-exports
- Fix rc re-exports
