+++ library/alloc/src/rc/rc.rs
@@ -1,245 +1,3 @@
-//! Single-threaded reference-counting pointers. 'Rc' stands for 'Reference
-//! Counted'.
-//!
-//! The type [`Rc<T>`][`Rc`] provides shared ownership of a value of type `T`,
-//! allocated in the heap. Invoking [`clone`][clone] on [`Rc`] produces a new
-//! pointer to the same allocation in the heap. When the last [`Rc`] pointer to a
-//! given allocation is destroyed, the value stored in that allocation (often
-//! referred to as "inner value") is also dropped.
-//!
-//! Shared references in Rust disallow mutation by default, and [`Rc`]
-//! is no exception: you cannot generally obtain a mutable reference to
-//! something inside an [`Rc`]. If you need mutability, put a [`Cell`]
-//! or [`RefCell`] inside the [`Rc`]; see [an example of mutability
-//! inside an `Rc`][mutability].
-//!
-//! [`Rc`] uses non-atomic reference counting. This means that overhead is very
-//! low, but an [`Rc`] cannot be sent between threads, and consequently [`Rc`]
-//! does not implement [`Send`][send]. As a result, the Rust compiler
-//! will check *at compile time* that you are not sending [`Rc`]s between
-//! threads. If you need multi-threaded, atomic reference counting, use
-//! [`sync::Arc`][arc].
-//!
-//! The [`downgrade`][downgrade] method can be used to create a non-owning
-//! [`Weak`] pointer. A [`Weak`] pointer can be [`upgrade`][upgrade]d
-//! to an [`Rc`], but this will return [`None`] if the value stored in the allocation has
-//! already been dropped. In other words, `Weak` pointers do not keep the value
-//! inside the allocation alive; however, they *do* keep the allocation
-//! (the backing store for the inner value) alive.
-//!
-//! A cycle between [`Rc`] pointers will never be deallocated. For this reason,
-//! [`Weak`] is used to break cycles. For example, a tree could have strong
-//! [`Rc`] pointers from parent nodes to children, and [`Weak`] pointers from
-//! children back to their parents.
-//!
-//! `Rc<T>` automatically dereferences to `T` (via the [`Deref`] trait),
-//! so you can call `T`'s methods on a value of type [`Rc<T>`][`Rc`]. To avoid name
-//! clashes with `T`'s methods, the methods of [`Rc<T>`][`Rc`] itself are associated
-//! functions, called using [fully qualified syntax]:
-//!
-//! ```
-//! use std::rc::Rc;
-//!
-//! let my_rc = Rc::new(());
-//! Rc::downgrade(&my_rc);
-//! ```
-//!
-//! `Rc<T>`'s implementations of traits like `Clone` may also be called using
-//! fully qualified syntax. Some people prefer to use fully qualified syntax,
-//! while others prefer using method-call syntax.
-//!
-//! ```
-//! use std::rc::Rc;
-//!
-//! let rc = Rc::new(());
-//! // Method-call syntax
-//! let rc2 = rc.clone();
-//! // Fully qualified syntax
-//! let rc3 = Rc::clone(&rc);
-//! ```
-//!
-//! [`Weak<T>`][`Weak`] does not auto-dereference to `T`, because the inner value may have
-//! already been dropped.
-//!
-//! # Cloning references
-//!
-//! Creating a new reference to the same allocation as an existing reference counted pointer
-//! is done using the `Clone` trait implemented for [`Rc<T>`][`Rc`] and [`Weak<T>`][`Weak`].
-//!
-//! ```
-//! use std::rc::Rc;
-//!
-//! let foo = Rc::new(vec![1.0, 2.0, 3.0]);
-//! // The two syntaxes below are equivalent.
-//! let a = foo.clone();
-//! let b = Rc::clone(&foo);
-//! // a and b both point to the same memory location as foo.
-//! ```
-//!
-//! The `Rc::clone(&from)` syntax is the most idiomatic because it conveys more explicitly
-//! the meaning of the code. In the example above, this syntax makes it easier to see that
-//! this code is creating a new reference rather than copying the whole content of foo.
-//!
-//! # Examples
-//!
-//! Consider a scenario where a set of `Gadget`s are owned by a given `Owner`.
-//! We want to have our `Gadget`s point to their `Owner`. We can't do this with
-//! unique ownership, because more than one gadget may belong to the same
-//! `Owner`. [`Rc`] allows us to share an `Owner` between multiple `Gadget`s,
-//! and have the `Owner` remain allocated as long as any `Gadget` points at it.
-//!
-//! ```
-//! use std::rc::Rc;
-//!
-//! struct Owner {
-//!     name: String,
-//!     // ...other fields
-//! }
-//!
-//! struct Gadget {
-//!     id: i32,
-//!     owner: Rc<Owner>,
-//!     // ...other fields
-//! }
-//!
-//! fn main() {
-//!     // Create a reference-counted `Owner`.
-//!     let gadget_owner: Rc<Owner> = Rc::new(
-//!         Owner {
-//!             name: "Gadget Man".to_string(),
-//!         }
-//!     );
-//!
-//!     // Create `Gadget`s belonging to `gadget_owner`. Cloning the `Rc<Owner>`
-//!     // gives us a new pointer to the same `Owner` allocation, incrementing
-//!     // the reference count in the process.
-//!     let gadget1 = Gadget {
-//!         id: 1,
-//!         owner: Rc::clone(&gadget_owner),
-//!     };
-//!     let gadget2 = Gadget {
-//!         id: 2,
-//!         owner: Rc::clone(&gadget_owner),
-//!     };
-//!
-//!     // Dispose of our local variable `gadget_owner`.
-//!     drop(gadget_owner);
-//!
-//!     // Despite dropping `gadget_owner`, we're still able to print out the name
-//!     // of the `Owner` of the `Gadget`s. This is because we've only dropped a
-//!     // single `Rc<Owner>`, not the `Owner` it points to. As long as there are
-//!     // other `Rc<Owner>` pointing at the same `Owner` allocation, it will remain
-//!     // live. The field projection `gadget1.owner.name` works because
-//!     // `Rc<Owner>` automatically dereferences to `Owner`.
-//!     println!("Gadget {} owned by {}", gadget1.id, gadget1.owner.name);
-//!     println!("Gadget {} owned by {}", gadget2.id, gadget2.owner.name);
-//!
-//!     // At the end of the function, `gadget1` and `gadget2` are destroyed, and
-//!     // with them the last counted references to our `Owner`. Gadget Man now
-//!     // gets destroyed as well.
-//! }
-//! ```
-//!
-//! If our requirements change, and we also need to be able to traverse from
-//! `Owner` to `Gadget`, we will run into problems. An [`Rc`] pointer from `Owner`
-//! to `Gadget` introduces a cycle. This means that their
-//! reference counts can never reach 0, and the allocation will never be destroyed:
-//! a memory leak. In order to get around this, we can use [`Weak`]
-//! pointers.
-//!
-//! Rust actually makes it somewhat difficult to produce this loop in the first
-//! place. In order to end up with two values that point at each other, one of
-//! them needs to be mutable. This is difficult because [`Rc`] enforces
-//! memory safety by only giving out shared references to the value it wraps,
-//! and these don't allow direct mutation. We need to wrap the part of the
-//! value we wish to mutate in a [`RefCell`], which provides *interior
-//! mutability*: a method to achieve mutability through a shared reference.
-//! [`RefCell`] enforces Rust's borrowing rules at runtime.
-//!
-//! ```
-//! use std::rc::Rc;
-//! use std::rc::Weak;
-//! use std::cell::RefCell;
-//!
-//! struct Owner {
-//!     name: String,
-//!     gadgets: RefCell<Vec<Weak<Gadget>>>,
-//!     // ...other fields
-//! }
-//!
-//! struct Gadget {
-//!     id: i32,
-//!     owner: Rc<Owner>,
-//!     // ...other fields
-//! }
-//!
-//! fn main() {
-//!     // Create a reference-counted `Owner`. Note that we've put the `Owner`'s
-//!     // vector of `Gadget`s inside a `RefCell` so that we can mutate it through
-//!     // a shared reference.
-//!     let gadget_owner: Rc<Owner> = Rc::new(
-//!         Owner {
-//!             name: "Gadget Man".to_string(),
-//!             gadgets: RefCell::new(vec![]),
-//!         }
-//!     );
-//!
-//!     // Create `Gadget`s belonging to `gadget_owner`, as before.
-//!     let gadget1 = Rc::new(
-//!         Gadget {
-//!             id: 1,
-//!             owner: Rc::clone(&gadget_owner),
-//!         }
-//!     );
-//!     let gadget2 = Rc::new(
-//!         Gadget {
-//!             id: 2,
-//!             owner: Rc::clone(&gadget_owner),
-//!         }
-//!     );
-//!
-//!     // Add the `Gadget`s to their `Owner`.
-//!     {
-//!         let mut gadgets = gadget_owner.gadgets.borrow_mut();
-//!         gadgets.push(Rc::downgrade(&gadget1));
-//!         gadgets.push(Rc::downgrade(&gadget2));
-//!
-//!         // `RefCell` dynamic borrow ends here.
-//!     }
-//!
-//!     // Iterate over our `Gadget`s, printing their details out.
-//!     for gadget_weak in gadget_owner.gadgets.borrow().iter() {
-//!
-//!         // `gadget_weak` is a `Weak<Gadget>`. Since `Weak` pointers can't
-//!         // guarantee the allocation still exists, we need to call
-//!         // `upgrade`, which returns an `Option<Rc<Gadget>>`.
-//!         //
-//!         // In this case we know the allocation still exists, so we simply
-//!         // `unwrap` the `Option`. In a more complicated program, you might
-//!         // need graceful error handling for a `None` result.
-//!
-//!         let gadget = gadget_weak.upgrade().unwrap();
-//!         println!("Gadget {} owned by {}", gadget.id, gadget.owner.name);
-//!     }
-//!
-//!     // At the end of the function, `gadget_owner`, `gadget1`, and `gadget2`
-//!     // are destroyed. There are now no strong (`Rc`) pointers to the
-//!     // gadgets, so they are destroyed. This zeroes the reference count on
-//!     // Gadget Man, so he gets destroyed as well.
-//! }
-//! ```
-//!
-//! [clone]: Clone::clone
-//! [`Cell`]: core::cell::Cell
-//! [`RefCell`]: core::cell::RefCell
-//! [send]: core::marker::Send
-//! [arc]: crate::sync::Arc
-//! [`Deref`]: core::ops::Deref
-//! [downgrade]: Rc::downgrade
-//! [upgrade]: Weak::upgrade
-//! [mutability]: core::cell#introducing-mutability-inside-of-something-immutable
-//! [fully qualified syntax]: https://doc.rust-lang.org/book/ch19-03-advanced-traits.html#fully-qualified-syntax-for-disambiguation-calling-methods-with-the-same-name
-
 #![stable(feature = "rust1", since = "1.0.0")]
 
 #[cfg(not(test))]
@@ -254,7 +12,6 @@ use core::cmp::Ordering;
 use core::convert::{From, TryFrom};
 use core::fmt;
 use core::hash::{Hash, Hasher};
-use core::intrinsics::abort;
 use core::iter;
 use core::marker::{self, PhantomData, Unpin, Unsize};
 use core::mem::{self, align_of_val_raw, forget, size_of_val};
@@ -263,24 +20,35 @@ use core::pin::Pin;
 use core::ptr::{self, NonNull};
 use core::slice::from_raw_parts_mut;
 
+use super::utils::{is_dangling, MarkerEq, RcInnerPtr};
 use crate::alloc::{
     box_free, handle_alloc_error, AllocError, Allocator, Global, Layout, WriteCloneIntoRaw,
 };
 use crate::borrow::{Cow, ToOwned};
+use crate::rc::Weak;
 use crate::string::String;
 use crate::vec::Vec;
-
-#[cfg(test)]
-mod tests;
 
 // This is repr(C) to future-proof against possible field-reordering, which
 // would interfere with otherwise safe [into|from]_raw() of transmutable
 // inner types.
 #[repr(C)]
-struct RcBox<T: ?Sized> {
-    strong: Cell<usize>,
-    weak: Cell<usize>,
-    value: T,
+pub(super) struct RcBox<T: ?Sized> {
+    pub(super) strong: Cell<usize>,
+    pub(super) weak: Cell<usize>,
+    pub(super) value: T,
+}
+
+impl<T: ?Sized> RcInnerPtr for RcBox<T> {
+    #[inline(always)]
+    fn weak_ref(&self) -> &Cell<usize> {
+        &self.weak
+    }
+
+    #[inline(always)]
+    fn strong_ref(&self) -> &Cell<usize> {
+        &self.strong
+    }
 }
 
 /// A single-threaded reference-counting pointer. 'Rc' stands for 'Reference
@@ -319,7 +87,7 @@ impl<T: ?Sized> Rc<T> {
         unsafe { self.ptr.as_ref() }
     }
 
-    fn from_inner(ptr: NonNull<RcBox<T>>) -> Self {
+    pub(super) fn from_inner(ptr: NonNull<RcBox<T>>) -> Self {
         Self { ptr, phantom: PhantomData }
     }
 
@@ -913,7 +681,7 @@ impl<T: ?Sized> Rc<T> {
     /// Returns `true` if there are no other `Rc` or [`Weak`] pointers to
     /// this allocation.
     #[inline]
-    fn is_unique(this: &Self) -> bool {
+    pub(super) fn is_unique(this: &Self) -> bool {
         Rc::weak_count(this) == 0 && Rc::strong_count(this) == 1
     }
 
@@ -1422,12 +1190,6 @@ impl<T: ?Sized + PartialEq> RcEqIdent<T> for Rc<T> {
     }
 }
 
-// Hack to allow specializing on `Eq` even though `Eq` has a method.
-#[rustc_unsafe_specialization_marker]
-pub(crate) trait MarkerEq: PartialEq<Self> {}
-
-impl<T: Eq> MarkerEq for T {}
-
 /// We're doing this specialization here, and not as a more general optimization on `&T`, because it
 /// would otherwise add a cost to all equality checks on refs. We assume that `Rc`s are used to
 /// store large values, that are slow to clone, but also heavy to check for equality, causing this
@@ -1854,431 +1616,6 @@ impl<T, I: iter::TrustedLen<Item = T>> ToRcSlice<T> for I {
     }
 }
 
-/// `Weak` is a version of [`Rc`] that holds a non-owning reference to the
-/// managed allocation. The allocation is accessed by calling [`upgrade`] on the `Weak`
-/// pointer, which returns an [`Option`]`<`[`Rc`]`<T>>`.
-///
-/// Since a `Weak` reference does not count towards ownership, it will not
-/// prevent the value stored in the allocation from being dropped, and `Weak` itself makes no
-/// guarantees about the value still being present. Thus it may return [`None`]
-/// when [`upgrade`]d. Note however that a `Weak` reference *does* prevent the allocation
-/// itself (the backing store) from being deallocated.
-///
-/// A `Weak` pointer is useful for keeping a temporary reference to the allocation
-/// managed by [`Rc`] without preventing its inner value from being dropped. It is also used to
-/// prevent circular references between [`Rc`] pointers, since mutual owning references
-/// would never allow either [`Rc`] to be dropped. For example, a tree could
-/// have strong [`Rc`] pointers from parent nodes to children, and `Weak`
-/// pointers from children back to their parents.
-///
-/// The typical way to obtain a `Weak` pointer is to call [`Rc::downgrade`].
-///
-/// [`upgrade`]: Weak::upgrade
-#[stable(feature = "rc_weak", since = "1.4.0")]
-pub struct Weak<T: ?Sized> {
-    // This is a `NonNull` to allow optimizing the size of this type in enums,
-    // but it is not necessarily a valid pointer.
-    // `Weak::new` sets this to `usize::MAX` so that it doesn’t need
-    // to allocate space on the heap.  That's not a value a real pointer
-    // will ever have because RcBox has alignment at least 2.
-    // This is only possible when `T: Sized`; unsized `T` never dangle.
-    ptr: NonNull<RcBox<T>>,
-}
-
-#[stable(feature = "rc_weak", since = "1.4.0")]
-impl<T: ?Sized> !marker::Send for Weak<T> {}
-#[stable(feature = "rc_weak", since = "1.4.0")]
-impl<T: ?Sized> !marker::Sync for Weak<T> {}
-
-#[unstable(feature = "coerce_unsized", issue = "27732")]
-impl<T: ?Sized + Unsize<U>, U: ?Sized> CoerceUnsized<Weak<U>> for Weak<T> {}
-
-#[unstable(feature = "dispatch_from_dyn", issue = "none")]
-impl<T: ?Sized + Unsize<U>, U: ?Sized> DispatchFromDyn<Weak<U>> for Weak<T> {}
-
-impl<T> Weak<T> {
-    /// Constructs a new `Weak<T>`, without allocating any memory.
-    /// Calling [`upgrade`] on the return value always gives [`None`].
-    ///
-    /// [`upgrade`]: Weak::upgrade
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::Weak;
-    ///
-    /// let empty: Weak<i64> = Weak::new();
-    /// assert!(empty.upgrade().is_none());
-    /// ```
-    #[stable(feature = "downgraded_weak", since = "1.10.0")]
-    pub fn new() -> Weak<T> {
-        Weak { ptr: NonNull::new(usize::MAX as *mut RcBox<T>).expect("MAX is not 0") }
-    }
-}
-
-pub(crate) fn is_dangling<T: ?Sized>(ptr: *mut T) -> bool {
-    let address = ptr as *mut () as usize;
-    address == usize::MAX
-}
-
-/// Helper type to allow accessing the reference counts without
-/// making any assertions about the data field.
-struct WeakInner<'a> {
-    weak: &'a Cell<usize>,
-    strong: &'a Cell<usize>,
-}
-
-impl<T: ?Sized> Weak<T> {
-    /// Returns a raw pointer to the object `T` pointed to by this `Weak<T>`.
-    ///
-    /// The pointer is valid only if there are some strong references. The pointer may be dangling,
-    /// unaligned or even [`null`] otherwise.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::Rc;
-    /// use std::ptr;
-    ///
-    /// let strong = Rc::new("hello".to_owned());
-    /// let weak = Rc::downgrade(&strong);
-    /// // Both point to the same object
-    /// assert!(ptr::eq(&*strong, weak.as_ptr()));
-    /// // The strong here keeps it alive, so we can still access the object.
-    /// assert_eq!("hello", unsafe { &*weak.as_ptr() });
-    ///
-    /// drop(strong);
-    /// // But not any more. We can do weak.as_ptr(), but accessing the pointer would lead to
-    /// // undefined behaviour.
-    /// // assert_eq!("hello", unsafe { &*weak.as_ptr() });
-    /// ```
-    ///
-    /// [`null`]: core::ptr::null
-    #[stable(feature = "rc_as_ptr", since = "1.45.0")]
-    pub fn as_ptr(&self) -> *const T {
-        let ptr: *mut RcBox<T> = NonNull::as_ptr(self.ptr);
-
-        if is_dangling(ptr) {
-            // If the pointer is dangling, we return the sentinel directly. This cannot be
-            // a valid payload address, as the payload is at least as aligned as RcBox (usize).
-            ptr as *const T
-        } else {
-            // SAFETY: if is_dangling returns false, then the pointer is dereferencable.
-            // The payload may be dropped at this point, and we have to maintain provenance,
-            // so use raw pointer manipulation.
-            unsafe { ptr::addr_of_mut!((*ptr).value) }
-        }
-    }
-
-    /// Consumes the `Weak<T>` and turns it into a raw pointer.
-    ///
-    /// This converts the weak pointer into a raw pointer, while still preserving the ownership of
-    /// one weak reference (the weak count is not modified by this operation). It can be turned
-    /// back into the `Weak<T>` with [`from_raw`].
-    ///
-    /// The same restrictions of accessing the target of the pointer as with
-    /// [`as_ptr`] apply.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::{Rc, Weak};
-    ///
-    /// let strong = Rc::new("hello".to_owned());
-    /// let weak = Rc::downgrade(&strong);
-    /// let raw = weak.into_raw();
-    ///
-    /// assert_eq!(1, Rc::weak_count(&strong));
-    /// assert_eq!("hello", unsafe { &*raw });
-    ///
-    /// drop(unsafe { Weak::from_raw(raw) });
-    /// assert_eq!(0, Rc::weak_count(&strong));
-    /// ```
-    ///
-    /// [`from_raw`]: Weak::from_raw
-    /// [`as_ptr`]: Weak::as_ptr
-    #[stable(feature = "weak_into_raw", since = "1.45.0")]
-    pub fn into_raw(self) -> *const T {
-        let result = self.as_ptr();
-        mem::forget(self);
-        result
-    }
-
-    /// Converts a raw pointer previously created by [`into_raw`] back into `Weak<T>`.
-    ///
-    /// This can be used to safely get a strong reference (by calling [`upgrade`]
-    /// later) or to deallocate the weak count by dropping the `Weak<T>`.
-    ///
-    /// It takes ownership of one weak reference (with the exception of pointers created by [`new`],
-    /// as these don't own anything; the method still works on them).
-    ///
-    /// # Safety
-    ///
-    /// The pointer must have originated from the [`into_raw`] and must still own its potential
-    /// weak reference.
-    ///
-    /// It is allowed for the strong count to be 0 at the time of calling this. Nevertheless, this
-    /// takes ownership of one weak reference currently represented as a raw pointer (the weak
-    /// count is not modified by this operation) and therefore it must be paired with a previous
-    /// call to [`into_raw`].
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::{Rc, Weak};
-    ///
-    /// let strong = Rc::new("hello".to_owned());
-    ///
-    /// let raw_1 = Rc::downgrade(&strong).into_raw();
-    /// let raw_2 = Rc::downgrade(&strong).into_raw();
-    ///
-    /// assert_eq!(2, Rc::weak_count(&strong));
-    ///
-    /// assert_eq!("hello", &*unsafe { Weak::from_raw(raw_1) }.upgrade().unwrap());
-    /// assert_eq!(1, Rc::weak_count(&strong));
-    ///
-    /// drop(strong);
-    ///
-    /// // Decrement the last weak count.
-    /// assert!(unsafe { Weak::from_raw(raw_2) }.upgrade().is_none());
-    /// ```
-    ///
-    /// [`into_raw`]: Weak::into_raw
-    /// [`upgrade`]: Weak::upgrade
-    /// [`new`]: Weak::new
-    #[stable(feature = "weak_into_raw", since = "1.45.0")]
-    pub unsafe fn from_raw(ptr: *const T) -> Self {
-        // See Weak::as_ptr for context on how the input pointer is derived.
-
-        let ptr = if is_dangling(ptr as *mut T) {
-            // This is a dangling Weak.
-            ptr as *mut RcBox<T>
-        } else {
-            // Otherwise, we're guaranteed the pointer came from a nondangling Weak.
-            // SAFETY: data_offset is safe to call, as ptr references a real (potentially dropped) T.
-            let offset = unsafe { data_offset(ptr) };
-            // Thus, we reverse the offset to get the whole RcBox.
-            // SAFETY: the pointer originated from a Weak, so this offset is safe.
-            unsafe { (ptr as *mut RcBox<T>).set_ptr_value((ptr as *mut u8).offset(-offset)) }
-        };
-
-        // SAFETY: we now have recovered the original Weak pointer, so can create the Weak.
-        Weak { ptr: unsafe { NonNull::new_unchecked(ptr) } }
-    }
-
-    /// Attempts to upgrade the `Weak` pointer to an [`Rc`], delaying
-    /// dropping of the inner value if successful.
-    ///
-    /// Returns [`None`] if the inner value has since been dropped.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::Rc;
-    ///
-    /// let five = Rc::new(5);
-    ///
-    /// let weak_five = Rc::downgrade(&five);
-    ///
-    /// let strong_five: Option<Rc<_>> = weak_five.upgrade();
-    /// assert!(strong_five.is_some());
-    ///
-    /// // Destroy all strong pointers.
-    /// drop(strong_five);
-    /// drop(five);
-    ///
-    /// assert!(weak_five.upgrade().is_none());
-    /// ```
-    #[stable(feature = "rc_weak", since = "1.4.0")]
-    pub fn upgrade(&self) -> Option<Rc<T>> {
-        let inner = self.inner()?;
-        if inner.strong() == 0 {
-            None
-        } else {
-            inner.inc_strong();
-            Some(Rc::from_inner(self.ptr))
-        }
-    }
-
-    /// Gets the number of strong (`Rc`) pointers pointing to this allocation.
-    ///
-    /// If `self` was created using [`Weak::new`], this will return 0.
-    #[stable(feature = "weak_counts", since = "1.41.0")]
-    pub fn strong_count(&self) -> usize {
-        if let Some(inner) = self.inner() { inner.strong() } else { 0 }
-    }
-
-    /// Gets the number of `Weak` pointers pointing to this allocation.
-    ///
-    /// If no strong pointers remain, this will return zero.
-    #[stable(feature = "weak_counts", since = "1.41.0")]
-    pub fn weak_count(&self) -> usize {
-        self.inner()
-            .map(|inner| {
-                if inner.strong() > 0 {
-                    inner.weak() - 1 // subtract the implicit weak ptr
-                } else {
-                    0
-                }
-            })
-            .unwrap_or(0)
-    }
-
-    /// Returns `None` when the pointer is dangling and there is no allocated `RcBox`,
-    /// (i.e., when this `Weak` was created by `Weak::new`).
-    #[inline]
-    fn inner(&self) -> Option<WeakInner<'_>> {
-        if is_dangling(self.ptr.as_ptr()) {
-            None
-        } else {
-            // We are careful to *not* create a reference covering the "data" field, as
-            // the field may be mutated concurrently (for example, if the last `Rc`
-            // is dropped, the data field will be dropped in-place).
-            Some(unsafe {
-                let ptr = self.ptr.as_ptr();
-                WeakInner { strong: &(*ptr).strong, weak: &(*ptr).weak }
-            })
-        }
-    }
-
-    /// Returns `true` if the two `Weak`s point to the same allocation (similar to
-    /// [`ptr::eq`]), or if both don't point to any allocation
-    /// (because they were created with `Weak::new()`).
-    ///
-    /// # Notes
-    ///
-    /// Since this compares pointers it means that `Weak::new()` will equal each
-    /// other, even though they don't point to any allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::Rc;
-    ///
-    /// let first_rc = Rc::new(5);
-    /// let first = Rc::downgrade(&first_rc);
-    /// let second = Rc::downgrade(&first_rc);
-    ///
-    /// assert!(first.ptr_eq(&second));
-    ///
-    /// let third_rc = Rc::new(5);
-    /// let third = Rc::downgrade(&third_rc);
-    ///
-    /// assert!(!first.ptr_eq(&third));
-    /// ```
-    ///
-    /// Comparing `Weak::new`.
-    ///
-    /// ```
-    /// use std::rc::{Rc, Weak};
-    ///
-    /// let first = Weak::new();
-    /// let second = Weak::new();
-    /// assert!(first.ptr_eq(&second));
-    ///
-    /// let third_rc = Rc::new(());
-    /// let third = Rc::downgrade(&third_rc);
-    /// assert!(!first.ptr_eq(&third));
-    /// ```
-    ///
-    /// [`ptr::eq`]: core::ptr::eq
-    #[inline]
-    #[stable(feature = "weak_ptr_eq", since = "1.39.0")]
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        self.ptr.as_ptr() == other.ptr.as_ptr()
-    }
-}
-
-#[stable(feature = "rc_weak", since = "1.4.0")]
-impl<T: ?Sized> Drop for Weak<T> {
-    /// Drops the `Weak` pointer.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::{Rc, Weak};
-    ///
-    /// struct Foo;
-    ///
-    /// impl Drop for Foo {
-    ///     fn drop(&mut self) {
-    ///         println!("dropped!");
-    ///     }
-    /// }
-    ///
-    /// let foo = Rc::new(Foo);
-    /// let weak_foo = Rc::downgrade(&foo);
-    /// let other_weak_foo = Weak::clone(&weak_foo);
-    ///
-    /// drop(weak_foo);   // Doesn't print anything
-    /// drop(foo);        // Prints "dropped!"
-    ///
-    /// assert!(other_weak_foo.upgrade().is_none());
-    /// ```
-    fn drop(&mut self) {
-        let inner = if let Some(inner) = self.inner() { inner } else { return };
-
-        inner.dec_weak();
-        // the weak count starts at 1, and will only go to zero if all
-        // the strong pointers have disappeared.
-        if inner.weak() == 0 {
-            unsafe {
-                Global.deallocate(self.ptr.cast(), Layout::for_value_raw(self.ptr.as_ptr()));
-            }
-        }
-    }
-}
-
-#[stable(feature = "rc_weak", since = "1.4.0")]
-impl<T: ?Sized> Clone for Weak<T> {
-    /// Makes a clone of the `Weak` pointer that points to the same allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::{Rc, Weak};
-    ///
-    /// let weak_five = Rc::downgrade(&Rc::new(5));
-    ///
-    /// let _ = Weak::clone(&weak_five);
-    /// ```
-    #[inline]
-    fn clone(&self) -> Weak<T> {
-        if let Some(inner) = self.inner() {
-            inner.inc_weak()
-        }
-        Weak { ptr: self.ptr }
-    }
-}
-
-#[stable(feature = "rc_weak", since = "1.4.0")]
-impl<T: ?Sized + fmt::Debug> fmt::Debug for Weak<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "(Weak)")
-    }
-}
-
-#[stable(feature = "downgraded_weak", since = "1.10.0")]
-impl<T> Default for Weak<T> {
-    /// Constructs a new `Weak<T>`, allocating memory for `T` without initializing
-    /// it. Calling [`upgrade`] on the return value always gives [`None`].
-    ///
-    /// [`None`]: Option
-    /// [`upgrade`]: Weak::upgrade
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::rc::Weak;
-    ///
-    /// let empty: Weak<i64> = Default::default();
-    /// assert!(empty.upgrade().is_none());
-    /// ```
-    fn default() -> Weak<T> {
-        Weak::new()
-    }
-}
-
 // NOTE: We checked_add here to deal with mem::forget safely. In particular
 // if you mem::forget Rcs (or Weaks), the ref-count can overflow, and then
 // you can free the allocation while outstanding Rcs (or Weaks) exist.
@@ -2287,84 +1624,6 @@ impl<T> Default for Weak<T> {
 //
 // This should have negligible overhead since you don't actually need to
 // clone these much in Rust thanks to ownership and move-semantics.
-
-#[doc(hidden)]
-trait RcInnerPtr {
-    fn weak_ref(&self) -> &Cell<usize>;
-    fn strong_ref(&self) -> &Cell<usize>;
-
-    #[inline]
-    fn strong(&self) -> usize {
-        self.strong_ref().get()
-    }
-
-    #[inline]
-    fn inc_strong(&self) {
-        let strong = self.strong();
-
-        // We want to abort on overflow instead of dropping the value.
-        // The reference count will never be zero when this is called;
-        // nevertheless, we insert an abort here to hint LLVM at
-        // an otherwise missed optimization.
-        if strong == 0 || strong == usize::MAX {
-            abort();
-        }
-        self.strong_ref().set(strong + 1);
-    }
-
-    #[inline]
-    fn dec_strong(&self) {
-        self.strong_ref().set(self.strong() - 1);
-    }
-
-    #[inline]
-    fn weak(&self) -> usize {
-        self.weak_ref().get()
-    }
-
-    #[inline]
-    fn inc_weak(&self) {
-        let weak = self.weak();
-
-        // We want to abort on overflow instead of dropping the value.
-        // The reference count will never be zero when this is called;
-        // nevertheless, we insert an abort here to hint LLVM at
-        // an otherwise missed optimization.
-        if weak == 0 || weak == usize::MAX {
-            abort();
-        }
-        self.weak_ref().set(weak + 1);
-    }
-
-    #[inline]
-    fn dec_weak(&self) {
-        self.weak_ref().set(self.weak() - 1);
-    }
-}
-
-impl<T: ?Sized> RcInnerPtr for RcBox<T> {
-    #[inline(always)]
-    fn weak_ref(&self) -> &Cell<usize> {
-        &self.weak
-    }
-
-    #[inline(always)]
-    fn strong_ref(&self) -> &Cell<usize> {
-        &self.strong
-    }
-}
-
-impl<'a> RcInnerPtr for WeakInner<'a> {
-    #[inline(always)]
-    fn weak_ref(&self) -> &Cell<usize> {
-        self.weak
-    }
-
-    #[inline(always)]
-    fn strong_ref(&self) -> &Cell<usize> {
-        self.strong
-    }
-}
 
 #[stable(feature = "rust1", since = "1.0.0")]
 impl<T: ?Sized> borrow::Borrow<T> for Rc<T> {
